Pack m_fields to uint32 in the nuclear and charge-dipole tables
Regenerate the tables with each M row's m_fields packed into one uint32
(component, ev index, rac x/y/z) instead of six bytes, and unpack it with
shifts/masks in the three engines (nuclear, charge-dipole matrix, charge-dipole
field). The NuclearTable / ChargeDipoleTable m_fields member becomes uint32.

Cuts tabulalib.so 13.7 → 11.4 MB (nuclear tables 2.4 → 1.5 MB, charge-dipole
7.5 → 5.7 MB); the nuclear and charge-dipole test suites pass unchanged.

+++ src/tabula/TabulaChargeDipoleTable.hpp
@@ -21,7 +21,8 @@ struct ChargeDipoleTable
     int ev_count, max_order, component_count, m_row_count;
     const std::uint8_t *ev_order, *ev_mu_power, *ev_goz_power, *ev_qx, *ev_qy, *ev_qz;
     const std::int8_t *component_bra_m, *component_ket_m;
-    const std::uint8_t *m_fields;
+    // packed per row: component (bits 0-6), ev index (7-18), rac x/y/z (19-22, 23-26, 27-30)
+    const std::uint32_t *m_fields;
     const std::uint16_t *m_coef_idx;
 };
